backend/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
annotated-types=0.6.0=pypi_0
anyio=4.2.0=pypi_0
assemblyai=0.20.2=pypi_0
bzip2=1.0.8=he774522_0
ca-certificates=2023.12.12=haa95532_0
certifi=2024.2.2=pypi_0
charset-normalizer=3.3.2=pypi_0
click=8.1.7=py312haa95532_0
colorama=0.4.6=py312haa95532_0
expat=2.5.0=hd77b12b_0
flask=2.2.5=py312haa95532_0
h11=0.14.0=pypi_0
httpcore=1.0.2=pypi_0
httpx=0.26.0=pypi_0
idna=3.6=pypi_0
itsdangerous=2.0.1=pyhd3eb1b0_0
jinja2=3.1.2=py312haa95532_0
libffi=3.4.4=hd77b12b_0
markupsafe=2.1.1=py312h2bbff1b_0
openssl=3.0.13=h2bbff1b_0
pip=23.3.1=py312haa95532_0
pydantic=2.6.1=pypi_0
pydantic-core=2.16.2=pypi_0
python=3.12.1=h1d929f7_0
python-dotenv=1.0.1=pypi_0
requests=2.31.0=pypi_0
setuptools=68.2.2=py312haa95532_0
sniffio=1.3.0=pypi_0
sqlite=3.41.2=h2bbff1b_0
tk=8.6.12=h2bbff1b_0
typing-extensions=4.9.0=pypi_0
tzdata=2023d=h04d1e81_0
urllib3=2.2.0=pypi_0
vc=14.2=h21ff451_1
vs2015_runtime=14.27.29016=h5e58377_2
websockets=12.0=pypi_0
werkzeug=2.3.8=py312haa95532_0
wheel=0.41.2=py312haa95532_0
xz=5.4.5=h8cc25b3_0
zlib=1.2.13=h8cc25b3_0
